Improvements:
- docker-compose.yml: added selinux labels; minor improvements.
- Updated 'install-docker-compose-fedora.sh'.
- Created 'set-monitoring.sh' to run it on-demand inside a container.
- Updated README.

+++ docker/scripts/install-docker-compose-fedora.sh
@@ -8,21 +8,39 @@ if [ "$(whoami)" != 'root' ]; then
     exit 1
 fi
 
-dnf remove docker* || true
+dnf remove -y docker* || true
 
-dnf install -y dnf-plugins-core
+dnf install -y dnf-plugins-core grubby
 
-dnf config-manager \
-    --add-repo \
-    https://download.docker.com/linux/fedora/docker-ce.repo
+dnf info moby-engine >/dev/null 2>&1
+MOBY_ENGINE_EXISTS=$?
+if [[ ${MOBY_ENGINE_EXISTS} == 0 ]]; then
+    dnf install moby-engine
+else
+    dnf config-manager \
+        --add-repo \
+        https://download.docker.com/linux/fedora/docker-ce.repo
 
-dnf config-manager --set-enabled docker-ce-edge
-dnf config-manager --set-enabled docker-ce-test
+#    dnf config-manager --set-disabled docker-ce-nightly
+#    dnf config-manager --set-disabled docker-ce-test
 
-dnf install -y docker-ce
+    dnf install -y docker-ce
+fi
 
 groupadd docker || true
 usermod -aG docker "$(logname)"
+
+FEDORA_VERSION=$(sed -nE 's/^VERSION_ID=(.*)$/\1/p' /etc/os-release)
+if [[ "${FEDORA_VERSION}" -ge 32 ]]; then
+    # Docker does not yet cooperate with the nftables backend.
+    firewall-cmd --permanent --zone=trusted --add-interface=docker0
+    firewall-cmd --permanent --zone=trusted --add-source=172.0.0.0/8
+    firewall-cmd --reload
+fi
+
+# Run SELinux in permissive mode.
+setenforce 0
+sed -i -E 's/^SELINUX=enforcing$/SELINUX=permissive/' /etc/selinux/config
 
 systemctl enable docker
 systemctl restart docker
@@ -33,5 +51,13 @@ curl -Ls "https://github.com/docker/compose/releases/download/$dockerComposeVers
     -o /usr/local/bin/docker-compose
 chmod +x /usr/local/bin/docker-compose
 
-echo "Docker successfully installed!!!
-Please log out and log in in order to use docker without 'sudo'."
+echo "Docker successfully installed!!!"
+
+if [[ "${FEDORA_VERSION}" -ge 31 ]]; then
+    # Docker does not yet support Control Group V2.
+    grubby --update-kernel=ALL --args="systemd.unified_cgroup_hierarchy=0"
+
+    echo "Please reboot in order for Cgroups backward compatibility to take effect."
+else
+    echo "Please log out and log in in order to use docker without 'sudo'."
+fi
